backend/requirements.txt
fastapi>=0.120.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
PyMuPDF>=1.22.0
pillow>=10.0.0
pix2tex>=0.1.0
torch>=2.0.0
